Suggestion for refactoring.
Start by putting ApplicationWindow in its own file.
